Move VehiclePlaceArgs to VehiclePlace.h
