+++ src/OpenLoco/src/GameCommands/GameCommands.cpp
@@ -70,6 +70,7 @@
 #include "Vehicles/VehiclePickup.h"
 #include "Vehicles/VehiclePickupAir.h"
 #include "Vehicles/VehiclePickupWater.h"
+#include "Vehicles/VehiclePlace.h"
 #include "Vehicles/VehicleRearrange.h"
 #include "Vehicles/VehicleRefit.h"
 #include "Vehicles/VehicleReverse.h"
