+++ src/OpenLoco/src/Windows/Vehicle.cpp
@@ -15,6 +15,7 @@
 #include "GameCommands/Vehicles/VehiclePickup.h"
 #include "GameCommands/Vehicles/VehiclePickupAir.h"
 #include "GameCommands/Vehicles/VehiclePickupWater.h"
+#include "GameCommands/Vehicles/VehiclePlace.h"
 #include "GameCommands/Vehicles/VehicleRearrange.h"
 #include "GameCommands/Vehicles/VehicleRefit.h"
 #include "GameCommands/Vehicles/VehicleReverse.h"
